+++ packages/format-library/src/text-color/index.jsx
@@ -6,15 +6,14 @@ import {
 	color as colorIcon,
 	textColor as textColorIcon,
 } from '@wordpress/icons';
-import { removeFormat } from '@wordpress/rich-text';
+import { toggleFormat } from '@wordpress/rich-text';
 import { default as InlineColorUI, getActiveColors } from './inline';
+import { getAvailableHighlightColors } from './utils';
 
 export const transparentValue = 'rgba(0, 0, 0, 0)';
 
 const name = 'core/text-color';
 const title = __( 'Highlight' );
-
-const EMPTY_ARRAY = [];
 
 function getComputedStyleProperty( element, property ) {
 	if ( ! element ) {
@@ -58,10 +57,26 @@ function TextColorEdit( {
 	activeAttributes,
 	contentRef,
 } ) {
-	const [ allowCustomControl, colors = EMPTY_ARRAY ] = useSettings(
+	const [
+		defaultColors,
+		themeColors,
+		customColors,
+		enableCustomColors,
+		enableDefaultColors,
+	] = useSettings(
+		'color.palette.default',
+		'color.palette.theme',
+		'color.palette.custom',
 		'color.custom',
-		'color.palette'
+		'color.defaultPalette'
 	);
+	const { colors, hasColorsToChoose } = getAvailableHighlightColors( {
+		themeColors,
+		defaultColors,
+		customColors,
+		enableCustomColors,
+		enableDefaultColors,
+	} );
 	const [ isAddingColor, setIsAddingColor ] = useState( false );
 	const colorIndicatorStyle = useMemo(
 		() =>
@@ -71,11 +86,6 @@ function TextColorEdit( {
 			),
 		[ contentRef, value, colors ]
 	);
-
-	const hasColorsToChoose = !! colors.length || allowCustomControl;
-	if ( ! hasColorsToChoose && ! isActive ) {
-		return null;
-	}
 
 	return (
 		<>
@@ -93,11 +103,16 @@ function TextColorEdit( {
 					/>
 				}
 				title={ title }
-				// If has no colors to choose but a color is active remove the color onClick.
 				onClick={
 					hasColorsToChoose
 						? () => setIsAddingColor( true )
-						: () => onChange( removeFormat( value, name ) )
+						: () =>
+								onChange(
+									toggleFormat( value, {
+										type: name,
+										title,
+									} )
+								)
 				}
 				role="menuitemcheckbox"
 			/>